Add cleanup of unused avatar resources to admin maintenance page - #1292
The user delete workflow needs to be corrected to include related resources like avatars, but for now this will provide a way to clean up any unneeded avatar file and database resources. The overall delete workflow issue will be more fully addressed in 2.6. The clean up code itself will be important going forward as part of conversion and upgrade processes.